dupla06/@gabrieRabelo, @BrayanArgradem: Resolve o exercicio Biblioteca

+++ src/main/java/com/thoughtworks/aceleradora/logica/IdentificadorDeArmstrong.java
@@ -8,10 +8,10 @@ public class IdentificadorDeArmstrong {
         String numerosString = Integer.toString(numero);
         String [] numeros = numerosString.split("");
         int tamanho = numeros.length;
-        int soma = 0;
+        double soma = 0;
         for(int i = 0; i < tamanho; i++) {
             int num = Integer.parseInt(numeros[i]);
-            soma = soma + (int)Math.pow(num, tamanho);
+            soma = soma + Math.pow(num, tamanho);
         }
         if(soma == numero){
             return "Este eh um numero de Armstrong!";
